Getting User by id

+++ src/main/java/WebbAppPVT15Grupp2/Services/UserService.java
@@ -47,7 +47,7 @@ public class UserService {
             logger.error("Unable to create. A User with name {} already exist", modUser.getUsername());
             return new ResponseEntity(HttpStatus.CONFLICT);
         }
-        repository.sproc_update_user(modUser.getId(), modUser.getUsername(), modUser.getPassword(),modUser.isActive(), modUser.getFacebook_login(), modUser.getFacebook_password(), modUser.getCurrentyouthcentre(), modUser.getRole());
+        repository.sproc_update_user(modUser.getId(), modUser.getUsername(), modUser.getPassword(), modUser.isActive(), modUser.getFacebook_login(), modUser.getFacebook_password(), modUser.getCurrentyouthcentre(), modUser.getRole());
 
         HttpHeaders headers = new HttpHeaders();
         headers.setLocation(ucBuilder.path("/addUser/{id}").buildAndExpand(modUser.getId()).toUri());
@@ -62,9 +62,9 @@ public class UserService {
     }
 
     @RequestMapping(value = "/user/{id}", method = RequestMethod.GET)
-    public ResponseEntity<?> getUserById(@PathVariable("id") long id) {
+    public ResponseEntity<?> getUserById(@PathVariable("id") int id) {
         logger.info("Fetching user by id {}", id);
-        User user = repository.findById((int) id).get();
+        User user = repository.findById(id).get();
 
         if (user == null) {
             logger.error("User with id {} not found.", id);
@@ -74,11 +74,7 @@ public class UserService {
 
 
     }
-
-
-
-
-
+    
 
     @GetMapping("/getPuser2")
     public List sproc_get_all_users() {
